Use new dialog for searching
