+++ DragonQuestinoWinDev/DragonQuestinoWinDev/DragonQuestinoWinDev.vcxproj
@@ -49,6 +49,7 @@
     <ClCompile Include="..\..\DragonQuestino\battle.c" />
     <ClCompile Include="..\..\DragonQuestino\enemy.c" />
     <ClCompile Include="..\..\DragonQuestino\game.c" />
+    <ClCompile Include="..\..\DragonQuestino\game_actions.c" />
     <ClCompile Include="..\..\DragonQuestino\game_data.c" />
     <ClCompile Include="..\..\DragonQuestino\game_input.c" />
     <ClCompile Include="..\..\DragonQuestino\game_items.c" />
